backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.44
psycopg[binary]==3.2.10
pydantic-settings==2.10.1
PyJWT==2.10.1
passlib==1.7.4
bcrypt==4.0.1
boto3==1.40.21
